Added new character selection screen

+++ src/Moxy/Moxy/Entities/Entity.cs
@@ -14,22 +14,21 @@ namespace Moxy.Entities
 		public Vector2 Location;
 		public Rectangle Collision;
 		public Vector2 CollisionCenter;
-		public float CollisionRadius;
+		public int CollisionRadius;
 		public Texture2D Texture;
 		public PlayerIndex playerIndex;
 		public EntityType EntityType;
 		public float Rotation;
-
 		public bool Alive = true;
 
 		public abstract void Update(GameTime gameTime);
 		public abstract void Draw(SpriteBatch batch);
+
 		public virtual void Draw(SpriteBatch batch, Rectangle ViewFrustrum)
 		{
 			if(ViewFrustrum.Contains(Collision))
 				Draw(batch);
 		}
-
 
 		public AnimationManager Animations;
 	}
